requirements3.txt
python-dotenv
faiss-cpu
langchain
langchain_community
langchain_huggingface
langchain_openai
nltk
numpy
pandas
gensim
transformers
scikit-learn
sentence-transformers 
pymupdf